Move fetchers' classes in an individual directory
