+++ setup.cfg
@@ -45,9 +45,9 @@ cip.formatters =
     glue = cloud_info_provider.formatters.glue:GLUE
     cmdb = cloud_info_provider.formatters.cmdb:CMDB
 cip.fetchers =
-    cloud = cloud_info_provider.core:CloudFetcher
-    storage = cloud_info_provider.core:StorageFetcher
-    compute = cloud_info_provider.core:ComputeFetcher
+    cloud = cloud_info_provider.fetchers.cloud:CloudFetcher
+    storage = cloud_info_provider.fetchers.storage:StorageFetcher
+    compute = cloud_info_provider.fetchers.compute:ComputeFetcher
 
 [egg_info]
 tag_build =
